fix edit driver documents
