+++ app/src/main/java/com/floadsbjm/floader/EditProfileDocumentsFragment.kt
@@ -1,17 +1,14 @@
 package com.floadsbjm.floader
 
-import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import com.floadsbjm.floader.databinding.FragmentEditVehicleBinding
+import com.floadsbjm.floader.databinding.FragmentEditProfileDocumentsBinding
 import com.floadsbjm.floader.network.BaseRepository
 import com.floadsbjm.floader.ui.base.BaseFragment
 import com.floadsbjm.floader.ui.viewmodels.ProfileViewModel
 
-class EditVehicleFragment :
-    BaseFragment<ProfileViewModel, FragmentEditVehicleBinding, BaseRepository>() {
+class EditProfileDocumentsFragment :
+    BaseFragment<ProfileViewModel, FragmentEditProfileDocumentsBinding, BaseRepository>() {
     override fun getViewModel(): Class<ProfileViewModel> {
         TODO("Not yet implemented")
     }
@@ -19,7 +16,9 @@ class EditVehicleFragment :
     override fun getFragmentBinding(
         inflater: LayoutInflater,
         container: ViewGroup?
-    ): FragmentEditVehicleBinding = FragmentEditVehicleBinding.inflate(inflater, container, false)
+    ): FragmentEditProfileDocumentsBinding = FragmentEditProfileDocumentsBinding.inflate(
+        inflater, container, false
+    )
 
     override fun getFragmentRepository(): BaseRepository {
         TODO("Not yet implemented")
