+++ app/src/main/java/com/floadsbjm/floader/ui/EditVehicleFragment.kt
@@ -1,27 +1,25 @@
-package com.floadsbjm.floader
+package com.floadsbjm.floader.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.floadsbjm.floader.databinding.FragmentResetPasswordOneBinding
+import com.floadsbjm.floader.databinding.FragmentEditVehicleBinding
 import com.floadsbjm.floader.network.BaseRepository
 import com.floadsbjm.floader.ui.base.BaseFragment
-import com.floadsbjm.floader.ui.viewmodels.AuthViewModel
+import com.floadsbjm.floader.ui.viewmodels.ProfileViewModel
 
-
-class ResetPasswordOneFragment :
-    BaseFragment<AuthViewModel, FragmentResetPasswordOneBinding, BaseRepository>() {
-    override fun getViewModel(): Class<AuthViewModel> {
+class EditVehicleFragment :
+    BaseFragment<ProfileViewModel, FragmentEditVehicleBinding, BaseRepository>() {
+    override fun getViewModel(): Class<ProfileViewModel> {
         TODO("Not yet implemented")
     }
 
     override fun getFragmentBinding(
         inflater: LayoutInflater,
         container: ViewGroup?
-    ): FragmentResetPasswordOneBinding =
-        FragmentResetPasswordOneBinding.inflate(inflater, container, false)
+    ): FragmentEditVehicleBinding = FragmentEditVehicleBinding.inflate(inflater, container, false)
 
     override fun getFragmentRepository(): BaseRepository {
         TODO("Not yet implemented")
