+++ app/src/main/res/drawable/dashed_rectangle.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <corners android:radius="@dimen/dimen5dp" />
+    <solid android:color="@color/white" />
+    <stroke
+        android:width="3dp"
+        android:color="@color/grey"
+        android:dashWidth="20px"
+        android:dashGap="10px" />
+</shape>
