+++ app/src/main/res/drawable/ic_income.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="28dp"
+    android:height="43dp"
+    android:viewportWidth="28"
+    android:viewportHeight="43">
+  <path
+      android:pathData="M19.7174,17.9311V13.515H21.2912C22.6444,13.515 23.8317,12.8511 24.3895,11.7819C24.9469,10.713 24.7386,9.5017 23.8447,8.6205L16.6046,1.4813C15.9591,0.8451 15.0285,0.48 14.0511,0.48C13.0736,0.48 12.143,0.8451 11.4976,1.4813L4.2567,8.6205C3.3632,9.5014 3.1545,10.713 3.7123,11.7819C4.2697,12.8511 5.457,13.515 6.8102,13.515H8.3837V17.9311C10.1499,17.3619 12.0591,17.0502 14.0511,17.0502C16.0427,17.0502 17.9515,17.3619 19.7174,17.9311Z"
+      android:fillColor="#575555"/>
+  <path
+      android:pathData="M27.3057,31.1886C27.3057,24.8523 21.3598,19.697 14.051,19.697C6.7422,19.697 0.7959,24.8523 0.7959,31.1886C0.7959,37.5252 6.7422,42.6805 14.051,42.6805C21.3598,42.6805 27.3057,37.5252 27.3057,31.1886ZM15.5773,36.6034V37.6952C15.5773,38.426 14.894,39.0188 14.051,39.0188C13.208,39.0188 12.5243,38.426 12.5243,37.6952V36.8713H11.4195C10.5765,36.8713 9.8928,36.2785 9.8928,35.5477C9.8928,34.8168 10.5765,34.2244 11.4195,34.2244H14.051C14.6604,34.2244 15.1558,33.6983 15.1558,33.0518C15.1558,32.4053 14.6604,31.8795 14.051,31.8795C11.7582,31.8795 9.8928,30.166 9.8928,28.06C9.8928,26.4493 10.985,25.069 12.5243,24.5088V23.417C12.5243,22.6858 13.208,22.0934 14.051,22.0934C14.894,22.0934 15.5773,22.6858 15.5773,23.417V24.2409H16.6825C17.5255,24.2409 18.2088,24.8333 18.2088,25.5645C18.2088,26.2954 17.5255,26.8878 16.6825,26.8878H14.051C13.4416,26.8878 12.9458,27.4139 12.9458,28.06C12.9458,28.7066 13.4416,29.2326 14.051,29.2326C16.3438,29.2326 18.2088,30.9458 18.2088,33.0518C18.2088,34.6629 17.117,36.0432 15.5773,36.6034Z"
+      android:fillColor="#575555"/>
+</vector>
