+++ app/src/main/res/drawable/ic_route.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="40dp"
+    android:height="39dp"
+    android:viewportWidth="40"
+    android:viewportHeight="39">
+  <path
+      android:pathData="M32.5281,17.1931C28.9089,17.1968 25.967,20.1619 25.9707,23.8035C25.9722,25.7928 26.9635,28.0604 28.9171,30.5444C30.3112,32.3174 31.6927,33.5497 31.7508,33.6012C32.2041,34.0036 32.8864,34.0008 33.3367,33.5997C33.3949,33.5482 34.7741,32.3129 36.1652,30.5377C37.6576,28.6323 39.0989,26.2019 39.0989,23.7908C39.0959,20.1562 36.1761,17.1931 32.5281,17.1931ZM30.7446,23.7587C30.743,22.77 31.5432,21.9686 32.5326,21.967C34.1253,21.967 34.9224,23.9005 33.8014,25.0216C32.6858,26.1396 30.7446,25.361 30.7446,23.7587Z"
+      android:fillColor="#575555"/>
+  <path
+      android:pathData="M6.6833,16.8918C7.1356,17.2933 7.8162,17.294 8.2691,16.8918C8.3273,16.8404 9.708,15.6066 11.1006,13.8328C13.0512,11.3467 14.0403,9.0783 14.0403,7.0882C14.0403,3.4474 11.0954,0.4846 7.4762,0.4846C3.857,0.4846 0.9121,3.4474 0.9121,7.0882C0.9121,9.0783 1.9012,11.3467 3.8518,13.8328C5.2444,15.6066 6.6251,16.8404 6.6833,16.8918ZM7.4762,5.2585C8.4631,5.2585 9.2664,6.0618 9.2664,7.0487C9.2664,8.0355 8.4631,8.8389 7.4762,8.8389C6.4894,8.8389 5.686,8.0355 5.686,7.0487C5.686,6.0618 6.4894,5.2585 7.4762,5.2585Z"
+      android:fillColor="#575555"/>
+  <path
+      android:pathData="M34.337,36.8839C34.337,38.5191 32.3091,39.3012 31.2138,38.0789H6.2822C3.3209,38.0789 0.9116,35.6696 0.9116,32.7083C0.9116,29.747 3.3209,27.3377 6.2822,27.3377H18.2542C19.8788,27.3377 21.2006,26.0159 21.2006,24.3913C21.2006,22.7667 19.8788,21.4449 18.2542,21.4449H8.8102C8.4499,21.8425 7.9703,22.0417 7.4757,22.0417C5.8887,22.0417 5.0871,20.1084 6.2084,18.9871C6.9455,18.2484 8.143,18.3025 8.8094,19.058H18.2542C21.1954,19.058 23.5875,21.4501 23.5875,24.3913C23.5875,27.3325 21.1954,29.7246 18.2542,29.7246H6.2822C4.6367,29.7246 3.2986,31.0628 3.2986,32.7083C3.2986,34.3538 4.6367,35.692 6.2822,35.692C33.1365,35.692 31.1744,35.727 31.2802,35.6211C32.3986,34.5002 34.3358,35.2875 34.337,36.8839Z"
+      android:fillColor="#575555"/>
+</vector>
