+++ app/src/main/res/drawable/rectangle_with_stroke.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
-    <solid android:color="@color/dark_white" />
+    <solid android:color="@color/cloud" />
     <stroke
         android:width="1dp"
         android:color="@color/dim_grey" />
