+++ app/src/main/res/layout/fragment_edit_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/dark_white"
+    android:background="@color/cloud"
     tools:context=".ui.EditProfileFragment">
 
     <View
