+++ app/src/main/res/layout/fragment_edit_vehicle.xml
@@ -4,18 +4,18 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/dark_white">
+    android:background="@color/cloud">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        tools:context=".EditVehicleFragment">
+        tools:context=".ui.EditVehicleFragment">
 
         <ImageView
             android:id="@+id/iv_arrow_back"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/margin24dp"
+            android:layout_marginStart="@dimen/margin16dp"
             android:layout_marginTop="@dimen/margin24dp"
             android:background="@drawable/ic_arrow"
             android:contentDescription="@string/iv_arrow"
