+++ app/src/main/res/layout/fragment_reset_password_one.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ResetPasswordOneFragment">
+    tools:context=".ui.ResetPasswordOneFragment">
 
     <ImageButton
         android:id="@+id/iv_arrow"
