+++ app/src/main/res/layout/fragment_reset_password_two.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ResetPasswordTwoFragment">
+    tools:context=".ui.ResetPasswordTwoFragment">
 
     <ImageButton
         android:id="@+id/iv_arrow"
