+++ app/src/main/res/values/colors.xml
@@ -8,6 +8,8 @@
     <color name="grey">#AFACAC</color>
     <color name="matterhorn">#575555</color>
     <color name="grey_2">#9C9B9B</color>
+    <color name="cloud">#ECECEC</color>
+    <color name="grey_3">#DADADA</color>
     <color name="dark_white">#ECE5E5</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
